+++ libraries/scenario/src/territory.rs
@@ -2,6 +2,6 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
 pub enum ControlledTerritory {
-    Place(usize),
+    Place(u32),
     Country([char; 3]),
 }
